Version 1.1:     resolved bugs;     added swipe to delete red background and white icon from left to right side;     added new icon for swipe to delete;     added several java classes for swipe to delete;     checkbox was changed on image button;

+++ app/src/main/java/com/pandacorp/taskui/ui/deleted_tasks/DeletedTasksFragment.java
@@ -13,6 +13,7 @@ import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.DividerItemDecoration;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -24,16 +25,17 @@ import com.pandacorp.taskui.DBHelper;
 import com.pandacorp.taskui.R;
 import com.pandacorp.taskui.ui.Adapter.CustomAdapter;
 import com.pandacorp.taskui.ui.Adapter.ListItem;
+import com.pandacorp.taskui.ui.Adapter.RecyclerItemTouchHelper;
 
 import java.util.ArrayList;
 
 import it.xabaras.android.recyclerview.swipedecorator.RecyclerViewSwipeDecorator;
 
-public class DeletedTasksFragment extends Fragment {
+public class DeletedTasksFragment extends Fragment implements RecyclerItemTouchHelper.RecyclerItemTouchHelperListener {
     private final String TAG = "MyLogs";
 
     private RecyclerView recyclerView;
-    private CustomAdapter adapter;
+    public CustomAdapter adapter;
     private ArrayList<String> itemList= new ArrayList<>();
     private ArrayList<ListItem> arrayItemList = new ArrayList<>();
 
@@ -100,6 +102,7 @@ public class DeletedTasksFragment extends Fragment {
         recyclerView.setHasFixedSize(false);
         recyclerView.setAdapter(adapter);
         recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
+
         enableSwipe();
         registerForContextMenu(recyclerView);
 
@@ -135,64 +138,52 @@ public class DeletedTasksFragment extends Fragment {
 
         }
     }
-
     private void enableSwipe() {
-        ItemTouchHelper.SimpleCallback simpleItemTouchCallback = new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT) {
-            @Override
-            public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, @NonNull RecyclerView.ViewHolder target) {
-                return false;
-            }
+        //Attached the ItemTouchHelper
+        recyclerView.addItemDecoration(new DividerItemDecoration(recyclerView.getContext(), DividerItemDecoration.VERTICAL));
 
-            @Override
-            public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
-                int position = viewHolder.getAdapterPosition();
-                final ListItem deletedModel = arrayItemList.get(position);
-                final int deletedPosition = position;
-                // deleting database item
-                final SQLiteDatabase WritableDatabase = dbHelper.getWritableDatabase();
-                WritableDatabase.delete(DBHelper.DELETED_TASKS_TABLE_NAME, DBHelper.KEY_TASK_TEXT + "=?", new String[]{deletedModel.getMainText()});
-                // showing snack bar with Undo option
-                Snackbar snackbar = Snackbar.make(getActivity().getWindow().getDecorView().getRootView(), getResources().getText(R.string.snackbar_removed), Snackbar.LENGTH_LONG);
-                snackbar.setAnchorView(R.id.fabs_constraintLayout_deleted);
-                snackbar.setAnimationMode(BaseTransientBottomBar.ANIMATION_MODE_SLIDE);
-                snackbar.setAction(getResources().getText(R.string.snackbar_undo), new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        // undo is selected, restore the deleted item
-                        adapter.restoreItem(deletedModel, deletedPosition);
-                        ContentValues contentValues = new ContentValues();
-                        contentValues.put(DBHelper.KEY_TASK_TEXT, deletedModel.getMainText());
-                        database.insert(DBHelper.DELETED_TASKS_TABLE_NAME, null, contentValues);
-                    }
-                });
-                snackbar.show();
-
-
-            }
-
-            @Override
-            public void onChildDraw(Canvas c, RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder, float dX, float dY, int actionState, boolean isCurrentlyActive) {
-                new RecyclerViewSwipeDecorator.Builder(getContext(), c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive)
-                        .addBackgroundColor(ContextCompat.getColor(getContext(), R.color.Red))
-                        .addActionIcon(R.drawable.ic_recyclerview_delete)
-                        .create()
-                        .decorate();
-
-                super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
-
-                super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
-            }
-
-
-        };
-        ItemTouchHelper itemTouchHelper = new ItemTouchHelper(simpleItemTouchCallback);
-        itemTouchHelper.attachToRecyclerView(recyclerView);
+        //Attached the ItemTouchHelper
+        ItemTouchHelper.SimpleCallback itemTouchHelperCallback = new RecyclerItemTouchHelper(0, ItemTouchHelper.LEFT, this);
+        new ItemTouchHelper(itemTouchHelperCallback).attachToRecyclerView(recyclerView);
     }
+
+
 
     @Override
     public void onDestroy() {
         super.onDestroy();
         cursor.close();
         database.close();
+    }
+
+    @Override
+    public void onSwiped(RecyclerView.ViewHolder viewHolder, int direction, int position) {
+        Log.d(TAG, "onSwiped: onSwiped");
+        if (viewHolder instanceof CustomAdapter.ViewHolder) {
+            final ListItem deletedModel = arrayItemList.get(position);
+            final int deletedPosition = position;
+            adapter.removeItem(position, deletedModel, DBHelper.DELETED_TASKS_TABLE_NAME);
+            // deleting database item
+            final SQLiteDatabase WritableDatabase = dbHelper.getWritableDatabase();
+            WritableDatabase.delete(DBHelper.DELETED_TASKS_TABLE_NAME, DBHelper.KEY_TASK_TEXT + "=?", new String[]{deletedModel.getMainText()});
+
+            // showing snack bar with Undo option
+            Snackbar snackbar = Snackbar.make(getActivity().getWindow().getDecorView().getRootView(), getResources().getText(R.string.snackbar_removed), Snackbar.LENGTH_LONG);
+            snackbar.setAnchorView(R.id.fabs_constraintLayout_deleted);
+            snackbar.setAnimationMode(BaseTransientBottomBar.ANIMATION_MODE_SLIDE);
+            snackbar.setAction(getResources().getText(R.string.snackbar_undo), new View.OnClickListener() {
+                @Override
+                public void onClick(View view) {
+                    // undo is selected, restore the deleted item
+                    adapter.restoreItem(deletedModel, deletedPosition);
+                    ContentValues contentValues = new ContentValues();
+                    contentValues.put(DBHelper.KEY_TASK_TEXT, deletedModel.getMainText());
+                    database.insert(DBHelper.COMPLETED_TASKS_TABLE_NAME, null, contentValues);
+                }
+            });
+            snackbar.show();
+        }
+
+
     }
 }
